Model modification for 2 table relations. Long tagId modificated in tag service.

+++ src/main/java/com/mls/Blog_API/service/ITagService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface ITagService {
     public List<Tag> listTags();
 
-    public Tag searchTagById(Integer tagId);
+    public Tag searchTagById(Long tagId);
 
     public List<Tag> searchTagByName (String tagName);
